+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `webfinger` command.
 - Sync `Update` and `Delete` activities in FEP-171b conversations.
 - Support SHA-512 content digests.
+- Support verification of signed requests with `Content-Digest` header.
 
 ### Changed