+++ apx_sdk/CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support SHA-512 content digests.
+- Support verification of signed requests with `Content-Digest` header.
+- Added `NoDigest` variant to `HttpSignatureVerificationError` enum.
 
 ### Changed